functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.0.0
sendgrid>=6.0.0
python-dotenv>=1.0.0